Apply auto-filtering to individual directives instead of an entire tag
Tags may contain multiple directives separated by ';', such as:

    [% IF foo; bar; ELSE; baz; END %]

Previously 'bar' and 'baz' were not auto-filtered by this module because
the directives IF and END appeared in the tag.  This made is easy for
variables to slip through unfiltered in compound tags.  The auto-filter
now translates the above examples into:

    [% IF foo; bar | html; ELSE; baz | html; END %]

Resolves [rt.cpan.org #90643].